added classes Migration, App and group classes Providers
