+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
     package="info.goodline.reshenie_plus">
 
     <application
-        android:name="android.support.multidex.MultiDexApplication"
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
