+++ app/src/main/java/info/goodline/reshenie_plus/extensions/App.kt
@@ -1,0 +1,18 @@
+package info.goodline.reshenie_plus
+
+import android.support.multidex.MultiDexApplication
+import info.goodline.reshenie_plus.models.Migration
+import io.realm.Realm
+import io.realm.RealmConfiguration
+
+class App : MultiDexApplication() {
+    override fun onCreate() {
+        super.onCreate()
+        Realm.init(this)
+
+        val config = RealmConfiguration.Builder()
+            .schemaVersion(Migration.SCHEMA_VERSION)
+            .migration(Migration()) // Migration to run instead of throwing an exception
+            .build()
+    }
+}
